+++ packages/fast-crud/src/components/crud/fs-form-item.vue
@@ -1,11 +1,19 @@
 <template>
-  <component
-    :is="$fsui.formItem.name"
-    class="fs-form-item"
-    :[$fsui.formItem.label]="item.title"
-    :[$fsui.formItem.prop]="item.key"
-    v-bind="item"
-  >
+  <component :is="$fsui.formItem.name" class="fs-form-item" :[$fsui.formItem.prop]="item.key" v-bind="item">
+    <template #label>
+      {{ item.label || item.title }}
+
+      <component
+        :is="$fsui.tooltip.name"
+        v-if="item.helper && computedHelperPosition === 'label'"
+        v-bind="computedHelperTooltip"
+      >
+        <template #[$fsui.tooltip.content]>
+          <fs-form-helper :helper="item.helper" :scope="buildItemScope(item)" />
+        </template>
+        <fs-icon class="fs-form-item-label-icon" :icon="$fsui.icons.question"></fs-icon>
+      </component>
+    </template>
     <fs-slot-render v-if="formSlot" :slots="formSlot" :scope="buildItemScope(item)" />
     <template v-else-if="item.component?.show !== false">
       <fs-render v-if="item.component?.render" :render-func="item.component.render" :scope="buildItemScope(item)" />
@@ -18,27 +26,22 @@
         @update:modelValue="updateModelValue"
       />
     </template>
-    <template v-if="item.helper">
-      <div class="fs-form-helper">
-        <template v-if="typeof item.helper === 'string'">{{ item.helper }}</template>
-        <template v-else-if="item.helper.render">
-          <fs-render :render-func="item.helper.render" :scope="buildItemScope(item)" />
-        </template>
-      </div>
+    <template v-if="item.helper && computedHelperPosition !== 'label'">
+      <fs-form-helper :helper="item.helper" :scope="buildItemScope(item)" />
     </template>
   </component>
 </template>
 <script>
-import traceUtil from "../../utils/util.trace";
 import FsRender from "../render/fs-render";
-import { ref } from "vue";
-
+import { ref, computed } from "vue";
+import FsFormHelper from "./fs-form-helper.vue";
+import _ from "lodash-es";
 /**
  * form-item组件封装
  */
 export default {
   name: "FsFormItem",
-  components: { FsRender },
+  components: { FsFormHelper, FsRender },
   props: {
     /**
      * 表单字段值(v-model)
@@ -64,11 +67,16 @@ export default {
     getContextFn: {
       type: Function,
       default: undefined
+    },
+    /**
+     * helper配置
+     */
+    helper: {
+      type: Object
     }
   },
   emits: ["update:modelValue"],
   setup(props, ctx) {
-    traceUtil.trace("fs-from-item");
     const componentRenderRef = ref();
     function buildItemScope(item) {
       const scope = props.getContextFn();
@@ -78,17 +86,35 @@ export default {
     function updateModelValue(value) {
       ctx.emit("update:modelValue", value);
     }
-    function getComponentRef() {
+    function getComponentRef(isAsync = false) {
+      if (isAsync) {
+        return componentRenderRef.value?.getTargetRefAsync();
+      }
       return componentRenderRef.value?.getTargetRef();
     }
+
+    const computedHelperPosition = computed(() => {
+      return props.item?.helper?.position || props.helper?.position;
+    });
+    const computedHelperTooltip = computed(() => {
+      return _.merge({}, props.item.helper?.tooltip, props.helper?.tooltip);
+    });
     return {
       updateModelValue,
       buildItemScope,
       getComponentRef,
-      componentRenderRef
+      componentRenderRef,
+      computedHelperPosition,
+      computedHelperTooltip
     };
   }
 };
 </script>
 
-<style lang="less"></style>
+<style lang="less">
+.fs-form-item {
+  .fs-form-item-label-icon {
+    margin: 0 2px;
+  }
+}
+</style>