+++ packages/fast-crud/src/d.ts/index.ts
@@ -1,0 +1,2 @@
+export * from "./expose";
+export * from "./types";